Fix tests meant for the advanced generator
For the most part, this was just replacing the URL from /generator to
/g. Some tests also don't work anymore with our new UX. I've removed
them for now and will write new ones covering these cases where
appropriate.
